Fix set_background function speed issue
Fix for issue https://github.com/openMLA/elegoo-mars-4-dlp-controller/issues. The problem was that integer type was incorrect (32 vs 8bit), and cast operations took a while. The 2nd issue was that for some reason the representation in memory was not very efficient; apparently taking the transpose of an intially "wrong" array shape is faster.